admin register basics
added the password interface for access to admin account registering. blade files for admin register created, as well as routing from the password interface to a new admin register page. issue with the validation not working in the RegisterController class but it is working in the LoginController class, so i have commented the code out in the RegisterController class for now and left it in the LoginController class