Admin creation and empty basket
Set isAdmin field to true when admin is created. Also when a user was initially made there was no basket. Also you could checkout with an empty basket. Now if basket is null or empty you are simply told the basket is empty rather than an error or allowing a checkout.